+++ app/TypiCMS/Modules/Categories/Presenters/ModulePresenter.php
@@ -1,8 +1,0 @@
-<?php
-namespace TypiCMS\Modules\Categories\Presenters;
-
-use TypiCMS\Presenters\Presenter;
-
-class ModulePresenter extends Presenter
-{
-}