+++ app/TypiCMS/Modules/Categories/lang/fr/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'catégories',
-    'categories' => 'catégorie|catégories',
-    'New'  => 'Nouvelle catégorie',
-    'Edit' => 'Modifier la catégorie',
-    'Back' => 'Retour à la liste des catégories',
-);