+++ app/TypiCMS/Modules/Contacts/Repositories/ContactInterface.php
@@ -1,9 +1,0 @@
-<?php
-namespace TypiCMS\Modules\Contacts\Repositories;
-
-use TypiCMS\Repositories\RepositoryInterface;
-
-interface ContactInterface extends RepositoryInterface
-{
-
-}