+++ app/TypiCMS/Modules/Contacts/lang/fr/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'contacts',
-    'contacts' => 'contact|contacts',
-    'New'  => 'Nouveau contact',
-    'Edit' => 'Modifier le contact',
-    'Back' => 'Retour à la liste des contacts',
-);