+++ app/TypiCMS/Modules/Dashboard/lang/en/global.php
@@ -1,6 +1,0 @@
-<?php
-return array(
-    'name' => 'Dashboard',
-    'Dashboard' => 'Dashboard',
-    'Welcome, :name!' => 'Welcome :name !',
-);