+++ app/TypiCMS/Modules/Groups/lang/fr/global.php
@@ -1,9 +1,0 @@
-<?php
-return array(
-    'name' => 'groupes',
-    'groups' => 'groupe|groupes',
-    'New'  => 'Nouveau groupe',
-    'Edit' => 'Modifier le groupe',
-    'Back' => 'Retour à la liste des groupes',
-    'Group permissions' => 'Permissions du groupe',
-);