+++ app/TypiCMS/Modules/History/Repositories/HistoryInterface.php
@@ -1,9 +1,0 @@
-<?php
-namespace TypiCMS\Modules\History\Repositories;
-
-use TypiCMS\Repositories\RepositoryInterface;
-
-interface HistoryInterface extends RepositoryInterface
-{
-
-}