+++ app/TypiCMS/Modules/Menulinks/lang/en/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'Menu links',
-    'menulinks' => 'Menu link|Menu links',
-    'New'  => 'New menu link',
-    'Edit' => 'Edit menu link',
-    'Back' => 'Back to menu links',
-);