+++ app/TypiCMS/Modules/News/Repositories/NewsInterface.php
@@ -1,9 +1,0 @@
-<?php
-namespace TypiCMS\Modules\News\Repositories;
-
-use TypiCMS\Repositories\RepositoryInterface;
-
-interface NewsInterface extends RepositoryInterface
-{
-
-}