+++ app/TypiCMS/Modules/News/lang/en/global.php
@@ -1,9 +1,0 @@
-<?php
-return array(
-    'name' => 'news',
-    'news' => 'news|news',
-    'New'  => 'New news',
-    'Edit' => 'Edit news',
-    'Back' => 'Back to news',
-    'Published on' => 'Published on',
-);