+++ app/TypiCMS/Modules/News/lang/fr/global.php
@@ -1,9 +1,0 @@
-<?php
-return array(
-    'name' => 'actualités',
-    'news' => 'actualité|actualités',
-    'New'  => 'Nouvelle actualité',
-    'Edit' => 'Modifier l’actualité',
-    'Back' => 'Retour à la liste des actualités',
-    'Published on' => 'Publié le',
-);