+++ app/TypiCMS/Modules/Partners/Presenters/ModulePresenter.php
@@ -1,9 +1,0 @@
-<?php
-namespace TypiCMS\Modules\Partners\Presenters;
-
-use TypiCMS\Presenters\Presenter;
-
-class ModulePresenter extends Presenter
-{
-
-}