+++ app/TypiCMS/Modules/Places/lang/en/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'places',
-    'places' => 'place|places',
-    'New'  => 'New place',
-    'Edit' => 'Edit place',
-    'Back' => 'Back to places',
-);