+++ app/TypiCMS/Modules/Places/lang/fr/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'adresses',
-    'places' => 'adresse|adresses',
-    'New'  => 'Nouvelle adresse',
-    'Edit' => 'Modifier l’adresse',
-    'Back' => 'Retour à la liste des adresses',
-);