+++ app/TypiCMS/Modules/Projects/lang/en/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'projects',
-    'projects' => 'project|projects',
-    'New'  => 'New project',
-    'Edit' => 'Edit project',
-    'Back' => 'Back to projects',
-);