+++ app/TypiCMS/Modules/Projects/lang/fr/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'projets',
-    'projects' => 'projet|projets',
-    'New'  => 'Nouveau projet',
-    'Edit' => 'Modifier le projet',
-    'Back' => 'Retour à la liste des projets',
-);