+++ app/TypiCMS/Modules/Tags/lang/en/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'tags',
-    'tags' => 'tag|tags',
-    'New'  => 'New tag',
-    'Edit' => 'Edit tag',
-    'Back' => 'Back to tags',
-);