+++ app/TypiCMS/Modules/Translations/lang/en/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'translations',
-    'translations' => 'translation|translations',
-    'New'  => 'New translation',
-    'Edit' => 'Edit translation',
-    'Back' => 'Back to translations',
-);