+++ app/TypiCMS/Modules/Translations/lang/fr/global.php
@@ -1,8 +1,0 @@
-<?php
-return array(
-    'name' => 'traductions',
-    'translations' => 'traduction|traductions',
-    'New'  => 'Nouvelle traduction',
-    'Edit' => 'Modifier la traduction',
-    'Back' => 'Retour à la liste des traductions',
-);